src/model/obj_detect_auto.py
import numpy as np

from tensorflow.keras.preprocessing import image
from tensorflow.keras.applications import EfficientNetB0
from tensorflow.keras.applications.efficientnet import preprocess_input


# загружаем модель для автоопределения
def get_obj_detect_auto():
    return EfficientNetB0(weights='imagenet')


# готовим изображение
def preprocess_image(img):
    img = img.resize((224, 224))

    img_array = image.img_to_array(img)
    img_array = np.expand_dims(img_array, axis=0)
    img_array = preprocess_input(img_array)

    return img_array
